Moved PubSub stuff to a trait.
